backend/requirements.txt
sanic==23.6.0
sanic-cors==2.2.0
tortoise-orm==0.20.0
pydantic==2.5.2
PyJWT==2.8.0 